# contracts/openapi/astro.json
{
  "components": {
    "schemas": {
      "AscendantPosition": {
        "additionalProperties": false,
        "properties": {
          "degree": {
            "exclusiveMaximum": 30.0,
            "format": "double",
            "minimum": 0.0,
            "title": "Degree",
            "type": "number"
          },
          "longitude": {
            "description": "Sidereal ecliptic longitude",
            "exclusiveMaximum": 360.0,
            "format": "double",
            "minimum": 0.0,
            "title": "Longitude",
            "type": "number"
          },
          "nakshatra": {
            "title": "Nakshatra",
            "type": "string"
          },
          "pada": {
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Pada",
            "type": "integer"
          },
          "sign": {
            "title": "Sign",
            "type": "string"
          },
          "sign_index": {
            "description": "0 = Aries",
            "maximum": 11.0,
            "minimum": 0.0,
            "title": "Sign Index",
            "type": "integer"
          }
        },
        "required": [
          "longitude",
          "sign",
          "sign_index",
          "degree",
          "nakshatra",
          "pada"
        ],
        "title": "AscendantPosition",
        "type": "object"
      },
      "BirthData": {
        "additionalProperties": false,
        "description": "Everything needed to compute a chart.\n\nThe instant is a single aware UTC timestamp rather than a local date,\na local time and a zone. Go owns timezone resolution \u2014 its stdlib\ncarries the full historical tzdata, including the 1942-45 Indian\nwartime offset \u2014 and resolves it once when the profile is created.\nRe-deriving it here would be a second implementation of the hardest\npart of the problem, free to disagree with the first.",
        "properties": {
          "ayanamsa": {
            "default": "lahiri",
            "enum": [
              "lahiri",
              "raman",
              "kp"
            ],
            "title": "Ayanamsa",
            "type": "string"
          },
          "house_system": {
            "const": "whole_sign",
            "default": "whole_sign",
            "title": "House System",
            "type": "string"
          },
          "latitude": {
            "format": "double",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "format": "double",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "time_accuracy": {
            "default": "exact",
            "enum": [
              "exact",
              "approximate",
              "unknown"
            ],
            "title": "Time Accuracy",
            "type": "string"
          },
          "utc_instant": {
            "description": "Birth instant in UTC. Must be timezone-aware.",
            "examples": [
              "1994-08-17T09:05:00Z"
            ],
            "format": "date-time",
            "title": "Utc Instant",
            "type": "string"
          }
        },
        "required": [
          "utc_instant",
          "latitude",
          "longitude"
        ],
        "title": "BirthData",
        "type": "object"
      },
...
      "ChartRequest": {
        "additionalProperties": false,
        "properties": {
          "birth": {
            "$ref": "#/components/schemas/BirthData"
          },
          "include_dasamsa": {
            "default": true,
            "title": "Include Dasamsa",
            "type": "boolean"
          },
          "include_dashas": {
            "default": true,
            "title": "Include Dashas",
            "type": "boolean"
          },
          "include_navamsa": {
            "default": true,
            "title": "Include Navamsa",
            "type": "boolean"
          },
          "include_yogas": {
            "default": true,
            "title": "Include Yogas",
            "type": "boolean"
          }
        },
        "required": [
          "birth"
        ],
        "title": "ChartRequest",
        "type": "object"
      },
      "ChartResponse": {
        "additionalProperties": false,
        "properties": {
          "ascendant": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AscendantPosition"
              },
              {
                "type": "null"
              }
            ]
          },
          "dasamsa": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DivisionalChart"
              },
              {
                "type": "null"
              }
            ]
          },
          "dashas": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DashaPeriodOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dashas"
          },
          "houses": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/HousePosition"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Houses"
          },
          "meta": {
            "$ref": "#/components/schemas/ChartMeta"
          },
          "navamsa": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DivisionalChart"
              },
              {
                "type": "null"
              }
            ]
          },
          "planets": {
            "items": {
              "$ref": "#/components/schemas/PlanetPosition"
            },
            "title": "Planets",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/ChartSummary"
          },
          "yogas": {
            "items": {
              "$ref": "#/components/schemas/YogaResult"
            },
            "title": "Yogas",
            "type": "array"
          }
        },
        "required": [
          "meta",
          "ascendant",
          "houses",
          "planets",
          "summary"
        ],
        "title": "ChartResponse",
        "type": "object"
      },
...
      "DashaPeriodOut": {
        "additionalProperties": false,
        "properties": {
          "children": {
            "items": {
              "$ref": "#/components/schemas/DashaPeriodOut"
            },
            "title": "Children",
            "type": "array"
          },
          "end": {
            "format": "date-time",
            "title": "End",
            "type": "string"
          },
          "level": {
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Level",
            "type": "integer"
          },
          "planet": {
            "title": "Planet",
            "type": "string"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "planet",
          "start",
          "end",
          "level"
        ],
        "title": "DashaPeriodOut",
        "type": "object"
      },
      "DashaRequest": {
        "additionalProperties": false,
        "description": "Dashas alone, for a caller that already has the Moon's position.",
        "properties": {
          "birth_instant": {
            "format": "date-time",
            "title": "Birth Instant",
            "type": "string"
          },
          "max_level": {
            "default": 3,
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Max Level",
            "type": "integer"
          },
          "moon_longitude": {
            "description": "Sidereal ecliptic longitude",
            "exclusiveMaximum": 360.0,
            "format": "double",
            "minimum": 0.0,
            "title": "Moon Longitude",
            "type": "number"
          }
        },
        "required": [
          "moon_longitude",
          "birth_instant"
        ],
        "title": "DashaRequest",
        "type": "object"
      },
      "DashaResponse": {
        "additionalProperties": false,
        "properties": {
          "balance_at_birth_days": {
            "format": "double",
            "title": "Balance At Birth Days",
            "type": "number"
          },
          "periods": {
            "items": {
              "$ref": "#/components/schemas/DashaPeriodOut"
            },
            "title": "Periods",
            "type": "array"
          }
        },
        "required": [
          "periods",
          "balance_at_birth_days"
        ],
        "title": "DashaResponse",
        "type": "object"
      },
      "DivisionalChart": {
        "additionalProperties": false,
        "description": "A D9 or D10. No yogas \u2014 they are read from the rasi.",
        "properties": {
          "ascendant": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AscendantPosition"
              },
              {
                "type": "null"
              }
            ]
          },
          "houses": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/HousePosition"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Houses"
          },
          "planets": {
            "items": {
              "$ref": "#/components/schemas/PlanetPosition"
            },
            "title": "Planets",
            "type": "array"
          }
        },
        "required": [
          "ascendant",
          "houses",
          "planets"
        ],
        "title": "DivisionalChart",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HousePosition": {
        "additionalProperties": false,
        "properties": {
          "house": {
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "House",
            "type": "integer"
          },
          "lord": {
            "title": "Lord",
            "type": "string"
          },
          "planets": {
            "items": {
              "type": "string"
            },
            "title": "Planets",
            "type": "array"
          },
          "sign": {
            "title": "Sign",
            "type": "string"
          },
          "sign_index": {
            "description": "0 = Aries",
            "maximum": 11.0,
            "minimum": 0.0,
            "title": "Sign Index",
            "type": "integer"
          }
        },
        "required": [
          "house",
          "sign",
          "sign_index",
          "lord"
        ],
        "title": "HousePosition",
        "type": "object"
      },
...
      "SadeSatiWindowsRequest": {
        "additionalProperties": false,
        "description": "All twelve windows at one instant.\n\nTwelve rather than one, because a Sade Sati window is a pure function\nof Saturn's motion and the natal Moon's SIGN \u2014 and there are only\ntwelve of those. api-service stores the twelve and serves every user\nfrom them, which is what lets the answer survive an astro outage: the\nalternative is a call to this service on the path of a question users\nask constantly.",
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "ayanamsa": {
            "default": "lahiri",
            "enum": [
              "lahiri",
              "raman",
              "kp"
            ],
            "title": "Ayanamsa",
            "type": "string"
          },
          "search_years": {
            "default": 40,
            "maximum": 80.0,
            "minimum": 10.0,
            "title": "Search Years",
            "type": "integer"
          }
        },
        "required": [
          "at"
        ],
        "title": "SadeSatiWindowsRequest",
        "type": "object"
      },
      "SadeSatiWindowsResponse": {
        "additionalProperties": false,
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "ayanamsa": {
            "title": "Ayanamsa",
            "type": "string"
          },
          "windows": {
            "items": {
              "$ref": "#/components/schemas/SadeSatiWindow"
            },
            "title": "Windows",
            "type": "array"
          }
        },
        "required": [
          "at",
          "ayanamsa",
          "windows"
        ],
        "title": "SadeSatiWindowsResponse",
        "type": "object"
      },
...
      "TransitRequest": {
        "additionalProperties": false,
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "ayanamsa": {
            "default": "lahiri",
            "enum": [
              "lahiri",
              "raman",
              "kp"
            ],
            "title": "Ayanamsa",
            "type": "string"
          },
          "natal_ascendant_sign": {
            "anyOf": [
              {
                "description": "0 = Aries",
                "maximum": 11.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Natal Ascendant Sign"
          },
          "natal_moon_sign": {
            "description": "0 = Aries",
            "maximum": 11.0,
            "minimum": 0.0,
            "title": "Natal Moon Sign",
            "type": "integer"
          }
        },
        "required": [
          "at",
          "natal_moon_sign"
        ],
        "title": "TransitRequest",
        "type": "object"
      },
      "TransitResponse": {
        "additionalProperties": false,
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "ayanamsa": {
            "title": "Ayanamsa",
            "type": "string"
          },
          "ayanamsa_value": {
            "format": "double",
            "title": "Ayanamsa Value",
            "type": "number"
          },
          "sade_sati": {
            "$ref": "#/components/schemas/SadeSatiResult"
          },
          "transits": {
            "items": {
              "$ref": "#/components/schemas/TransitPosition"
            },
            "title": "Transits",
            "type": "array"
          }
        },
        "required": [
          "at",
          "ayanamsa",
          "ayanamsa_value",
          "transits",
          "sade_sati"
        ],
        "title": "TransitResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Deterministic Vedic astrology computation. Internal only.",
    "title": "astro-service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/charts/compute": {
      "post": {
        "description": "Birth data in, full chart out.\n\nIdempotent and stateless. Nothing is written anywhere \u2014 `api-service`\nowns every byte of persistence, and this service is a pure function\nthat happens to be reachable over HTTP.",
        "operationId": "compute_chart_v1_charts_compute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Chart",
        "tags": [
          "charts"
        ]
      }
    },
    "/v1/dashas/compute": {
      "post": {
        "description": "The Vimshottari tree from a Moon longitude.\n\nSeparate from the chart endpoint so `api-service` can recompute a\ntree \u2014 after a birth-time correction, say \u2014 without recomputing every\nplanet.",
        "operationId": "compute_dashas_v1_dashas_compute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Dashas",
        "tags": [
          "charts"
        ]
      }
    },
    "/v1/transits/compute": {
      "post": {
        "description": "Current positions against a natal chart, plus Sade Sati.",
        "operationId": "compute_transit_v1_transits_compute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Transit",
        "tags": [
          "charts"
        ]
      }
    },
    "/v1/transits/sade-sati/windows": {
      "post": {
        "description": "Every natal Moon sign's Sade Sati window at one instant.\n\nTwelve answers in one call, because a window depends only on Saturn's\nmotion and the natal Moon's SIGN \u2014 and there are twelve of those.\napi-service stores the twelve and serves every user from them, so the\nquestion \"when does this end\" keeps having an answer while this\nservice is unreachable.\n\nThe scan is the expensive part and it is shared: `sade_sati_window`\nre-finds Saturn's ingresses per sign today, which is wasteful and\ncorrect. Sharing the ingress scan across the twelve is the obvious\noptimisation and is deliberately not done here \u2014 this runs once every\nsix hours on a worker, and a faster version of a function with three\nsubtle cases in it is not worth the risk of getting one wrong.",
        "operationId": "compute_sade_sati_windows_v1_transits_sade_sati_windows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SadeSatiWindowsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SadeSatiWindowsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Sade Sati Windows",
        "tags": [
          "charts"
        ]
      }
    }
  }
}